Added maven lein connector
