+++ pom.xml
@@ -45,7 +45,7 @@
     <dependency>
       <groupId>org.slf4j</groupId>
       <artifactId>slf4j-api</artifactId>
-      <version>1.7.6</version>
+      <version>1.7.10</version>
     </dependency>
     <dependency>
       <groupId>ch.qos.logback</groupId>
@@ -208,6 +208,19 @@
               <Eclipse-SourceBundle>${groupId}.${artifactId};version="${parsedVersion.osgiVersion}";roots:="."</Eclipse-SourceBundle>
             </manifestEntries>
           </archive>
+        </configuration>
+      </plugin>
+
+      <plugin>
+        <groupId>com.theoryinpractise</groupId>
+        <artifactId>clojure-maven-plugin</artifactId>
+        <version>1.3.20</version>
+        <extensions>true</extensions>
+        <configuration>
+          <runWithTests>false</runWithTests>
+          <testSourceDirectories>
+            <testSourceDirectory>src/test/clojure</testSourceDirectory>
+          </testSourceDirectories>
         </configuration>
       </plugin>
