+++ src/main/java/me/moocar/logbackgelf/GelfUDPAppender.java
@@ -1,0 +1,7 @@
+package me.moocar.logbackgelf;
+
+/**
+ * Created by anthony on 3/8/15.
+ */
+public class GelfUDPAppender {
+}
