+++ src/main/java/me/moocar/logbackgelf/GelfUDPOutputStream.java
@@ -1,0 +1,7 @@
+package me.moocar.logbackgelf;
+
+/**
+ * Created by anthony on 3/8/15.
+ */
+public class GelfUDPOutputStream {
+}
